Fix `git-clean-check` Makefile target
The `git_dirty` variable was not being set in the Makefile, so the
`git-clean-check` was incorrectly always returning a success exit code.
Set the `git_dirty` variable so that the make target works as expected.

Also set `--no-pager` in the Git commands to better handle environments
where the Git configuration uses the pager for output, which requires
interactive input to close the editor with the Git command output.

Signed-off-by: Ryan Turner <turner@uber.com>